requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.6.1
email-validator==2.2.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
PyJWT==2.9.0
httpx==0.27.2
python-multipart==0.0.9